Consolidate game.sources and settings.sources into source
