+++ data/src/commonMain/kotlin/com/jayasuryat/data/di/DataSourceModule.kt
@@ -16,10 +16,10 @@
 package com.jayasuryat.data.di
 
 import com.jayasuryat.data.MinesweeperDatabase
-import com.jayasuryat.data.game.sources.definition.GameDataSource
-import com.jayasuryat.data.game.sources.impl.GameDataSourceImpl
-import com.jayasuryat.data.settings.sources.definitions.UserPreferences
-import com.jayasuryat.data.settings.sources.impl.UserPreferencesImpl
+import com.jayasuryat.data.source.definition.GameDataSource
+import com.jayasuryat.data.source.impl.GameDataSourceImpl
+import com.jayasuryat.data.source.definition.UserPreferences
+import com.jayasuryat.data.source.impl.UserPreferencesImpl
 import com.jayasuryat.data.store.DataStore
 import kotlinx.serialization.json.Json
 import org.koin.core.module.Module
